updated version C

+++ eclipse-workspace/REST_WS_practica_psp/src/main/java/com/pspro/rest/HospitalServiceController.java
@@ -17,7 +17,7 @@ import com.pspro.rest.model.Hospital;
 /**
  * Clase  encargada de procesar peticiones. Tiene recursos de nivel 1.
  * @author: Carlos Jiménez
- * @version: 10/02/2021/B
+ * @version: 01/03/2021/C
  *
  */
 @RestController
@@ -40,34 +40,55 @@ public class HospitalServiceController {
 		hospRepo.put(hosp2.getId(), hosp2);
 	}
 
+	// Método que borra el hospital introducido:
 	@DeleteMapping("/hospitales/{id}")
 	public ResponseEntity<Object> delete(@PathVariable("id") String id) {
-		hospRepo.remove(id);
-		return new ResponseEntity<>("Hospital se ha eliminado con éxito", HttpStatus.OK);
+		if(hospRepo.get(id) != null) {
+			hospRepo.remove(id);
+			return new ResponseEntity<>("El Hospital se ha eliminado con éxito", HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>("El Hospital indicado no existe", HttpStatus.OK);
+		}
 	}
 
+	// Actualiza el hospital introducido con los diferentes campos deseados:
 	@PutMapping("/hospitales/{id}")
 	public ResponseEntity<Object> updateHospital(@PathVariable("id") String id, @RequestBody Hospital hosp) {
-		hospRepo.remove(id);
-		hosp.setId(id);
-		hospRepo.put(id, hosp);
-		return new ResponseEntity<>("Hospital se ha actualizado con éxito", HttpStatus.OK);
+		if(hospRepo.get(id) != null) {
+			hospRepo.remove(id);
+			hosp.setId(id);
+			hospRepo.put(id, hosp);
+			return new ResponseEntity<>("El Hospital se ha actualizado con éxito", HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>("El Hospital indicado no existe", HttpStatus.OK);
+		}
 	}
-
+	
+	// Crea el hospital introducido con los diferentes campos deseados:
 	@PostMapping("/hospitales")
 	public ResponseEntity<Object> createEspecialidad(@RequestBody Hospital hosp) {
-		hospRepo.put(hosp.getId(), hosp);
-		return new ResponseEntity<>("Hospital se ha creado con éxito", HttpStatus.CREATED);
+		if (hosp.getId() != null && hospRepo.get(hosp.getId()) == null) {
+			hospRepo.put(hosp.getId(), hosp);
+			return new ResponseEntity<>("El Hospital se ha creado con éxito", HttpStatus.CREATED);
+		} else {
+			return new ResponseEntity<>("El Hospital no se pudo crear", HttpStatus.CREATED);
+		}	
 	}
 
+	// Muestra todos los hospitales:
 	@GetMapping("/hospitales")
 	public ResponseEntity<Object> getHospitales() {
 		return new ResponseEntity<>(hospRepo.values(), HttpStatus.OK);
 	}
-
+	
+	// Muestra el hospital introducido:
 	@GetMapping("/hospitales/{id}")
 	public ResponseEntity<Object> getHospitalById(@PathVariable("id") String id) {
-		return new ResponseEntity<>(hospRepo.get(id), HttpStatus.OK);
+		if(hospRepo.get(id) != null) {
+			return new ResponseEntity<>(hospRepo.get(id), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>("El Hospital indicado no existe", HttpStatus.OK);
+		}
 	}
 
 }
